Change COPYING file to BSD license and remove LICENSE file
The project was intended to be under BSD license since the beginning,
and a LICENSE file was provided accordingly. However, automake
generated a COPYING file with the GPLv3 by default. The COPYING file
now contains the BSD license, and the LICENSE file is removed.

Reported by Török Edwin,
https://github.com/thierry-martinez/stdcompat/issues/5